add manabar

+++ dungeon/src/contrib/skill/Resource.java
@@ -14,7 +14,7 @@ import java.util.function.Function;
 public enum Resource {
   HP(
       entity ->
-          entity.fetch(HealthComponent.class).map(HealthComponent::currentHealthpoints).orElse(0),
+              Float.valueOf(entity.fetch(HealthComponent.class).map(HealthComponent::currentHealthpoints).orElse(0)),
       (entity, amount) ->
           entity
               .fetch(HealthComponent.class)
@@ -22,32 +22,32 @@ public enum Resource {
               .receiveHit(new Damage(amount, DamageType.SKILL, entity))),
   ARROW(
       entity ->
-          entity
-              .fetch(InventoryComponent.class)
-              .map(ic -> ic.count(ItemWoodenArrow.class))
-              .orElse(0),
+              Float.valueOf(entity
+                  .fetch(InventoryComponent.class)
+                  .map(ic -> ic.count(ItemWoodenArrow.class))
+                  .orElse(0)),
       (entity, amount) ->
           entity
               .fetch(InventoryComponent.class)
               .orElseThrow()
               .remove(ItemWoodenArrow.class, amount)),
   MANA(
-      entity -> entity.fetch(ManaComponent.class).map(ManaComponent::getCurrentAmount).orElse(0),
+      entity -> entity.fetch(ManaComponent.class).map(ManaComponent::getCurrentAmount).orElse((float) 0),
       (entity, amount) -> entity.fetch(ManaComponent.class).orElseThrow().consume(amount)),
   ENERGY(
       entity ->
-          entity.fetch(EnergyComponent.class).map(EnergyComponent::getCurrentAmount).orElse(0),
+          entity.fetch(EnergyComponent.class).map(EnergyComponent::getCurrentAmount).orElse((float) 0),
       (entity, amount) -> entity.fetch(EnergyComponent.class).orElseThrow().consume(amount));
 
-  private final Function<Entity, Integer> supplier;
+  private final Function<Entity, Float> supplier;
   private final BiConsumer<Entity, Integer> consume;
 
-  Resource(Function<Entity, Integer> supplier, BiConsumer<Entity, Integer> consume) {
+  Resource(Function<Entity, Float> supplier, BiConsumer<Entity, Integer> consume) {
     this.supplier = supplier;
     this.consume = consume;
   }
 
-  public int apply(Entity entity) {
+  public float apply(Entity entity) {
     return supplier.apply(entity);
   }
